+++ src/main/java/de/muspellheim/commons/domain/AuthorizationException.java
@@ -1,0 +1,18 @@
+/*
+ * Muspellheim Commons
+ * Copyright (c) 2019 Falko Schumann
+ */
+
+package de.muspellheim.commons.domain;
+
+/** Thrown by a repository when a user login is not possible. */
+public class AuthorizationException extends RepositoryException {
+
+  public AuthorizationException(String message) {
+    super(message);
+  }
+
+  public AuthorizationException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}