+++ src/main/java/de/muspellheim/commons/domain/ConnectionException.java
@@ -1,0 +1,18 @@
+/*
+ * Muspellheim Commons
+ * Copyright (c) 2019 Falko Schumann
+ */
+
+package de.muspellheim.commons.domain;
+
+/** Thrown by a repository when a connection to data store is not possible. */
+public class ConnectionException extends RepositoryException {
+
+  public ConnectionException(String message) {
+    super(message);
+  }
+
+  public ConnectionException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}