+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+*   Repository exceptions for connection and authorization.
 ## [v2.1.0] - 2019-11-20
 
 ### Added